Nevermind we might be okay
Rooms 2 and 3 are now fully working so now it should all be good hopefully to fix any bugs , add the audio queues and the final couple of screens

+++ Its-a-Whiskey-Buisness/Assets/Scripts/BackgroundLogic/GenericRoom.cs
@@ -458,9 +458,9 @@ public class GenericRoom : MonoBehaviour
     {
         if (RoomNum == sender.GetComponent<ClickDialObject>().Room_num /*|| RoomNum == sender.GetComponent<GenericRoomManager>().Roomnum*/)
         {
-            if (data is float)
+            if (data is int)
             {
-                Temperature = (float)data;
+                Temperature = (int)data;
                 SetTemperature.Raise(this, Temperature);
                 //Raise new temperature
             }
